fix: util 다운로드 함수 수정
