--- src/utils/download.ts
@@ -1,8 +1,0 @@
-export function downloadImage(url: string, filename: string) {
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = filename;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-}
